+++ apps/web/app/[locale]/website-editor/[websiteId]/pages/[pageId]/PageBuilder.tsx
@@ -343,6 +343,8 @@ export function PageBuilder({
     description: string;
     focusKeyword: string;
     image: string;
+    pixelEvent: string;
+    headCode: string;
   };
   domain: string;
   ogImageUrl?: string;